# PolyclinicLab/Polyclinic.Client.Wasm/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Polyclinic API",
    "version": "v1"
  },
  "paths": {
...
    "/api/appointments": {
      "get": {
        "tags": ["Appointments"],
        "operationId": "GetAppointments",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Appointments"],
        "operationId": "CreateAppointment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{id}": {
      "get": {
        "tags": ["Appointments"],
        "operationId": "GetAppointment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Appointments"],
        "operationId": "UpdateAppointment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Appointments"],
        "operationId": "DeleteAppointment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/analytics/experienced-doctors": {
      "get": {
        "tags": ["Analytics"],
        "operationId": "GetExperiencedDoctors",
        "parameters": [
          {
            "name": "minExperience",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/patients-by-doctor/{doctorId}": {
      "get": {
        "tags": ["Analytics"],
        "operationId": "GetPatientsByDoctor",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/repeated-appointments-count": {
      "get": {
        "tags": ["Analytics"],
        "operationId": "GetRepeatedAppointments",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepeatedAppointmentsAnalyticsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/patients-older-than-with-multiple-doctors": {
      "get": {
        "tags": ["Analytics"],
        "operationId": "GetPatientsOlderThanWithMultipleDoctors",
        "parameters": [
          {
            "name": "age",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientAnalyticsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/appointments-in-room/{room}": {
      "get": {
        "tags": ["Analytics"],
        "operationId": "GetAppointmentsInRoom",
        "parameters": [
          {
            "name": "room",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "passport": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "bloodType": {
            "type": "string",
            "nullable": true
          },
          "rhFactor": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "DoctorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "passport": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "birthYear": {
            "type": "integer",
            "format": "int32"
          },
          "specialization": {
            "type": "string",
            "nullable": true
          },
          "experience": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AppointmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "room": {
            "type": "integer",
            "format": "int32"
          },
          "isRepeated": {
            "type": "boolean"
          },
          "doctorName": {
            "type": "string",
            "nullable": true
          },
          "patientName": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateUpdateAppointmentDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "room": {
            "type": "integer",
            "format": "int32"
          },
          "isRepeated": {
            "type": "boolean"
          },
          "doctorId": {
            "type": "integer",
            "format": "int32"
          },
          "patientId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RepeatedAppointmentsAnalyticsDto": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppointmentDto"
            }
          }
        }
      },
      "PatientAnalyticsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "doctorCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}